pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "datamigrate-qa"
version = "0.1.0"
description = "Automated data migration QA tool"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "psycopg2-binary>=2.9",
    "jinja2>=3.1",
    "rich>=13.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
snowflake = ["snowflake-connector-python>=3.0"]
oracle = ["oracledb>=2.0"]
ai = ["anthropic>=0.40.0"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "testcontainers[postgres]>=4.0",
    "mypy>=1.10",
    "ruff>=0.4",
]

[project.scripts]
datamigrate-qa = "datamigrate_qa.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/datamigrate_qa"]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
